requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-rest-framework==0.1.0
djangorestframework==3.7.3  # via django-rest-framework
psycopg2==2.7.3.2
